Creating CoreEntity
