+++ tests/Domain/Entity/BrandTest.php
@@ -125,7 +125,7 @@ class BrandTest extends TestCase
 
     /**
      * @expectedException \BBC\ProgrammesPagesService\Domain\Exception\DataNotFetchedException
-     * @expectedExceptionMessage Could not get MasterBrand of Programme "p01m5mss" as it was not fetched
+     * @expectedExceptionMessage Could not get MasterBrand of CoreEntity "p01m5mss" as it was not fetched
      */
     public function testRequestingUnfetchedMasterBrandThrowsException()
     {
@@ -164,7 +164,7 @@ class BrandTest extends TestCase
 
     /**
      * @expectedException \BBC\ProgrammesPagesService\Domain\Exception\DataNotFetchedException
-     * @expectedExceptionMessage Could not get options of Programme "p01m5mss"as the full hierarchy was not fetched
+     * @expectedExceptionMessage Could not get options of CoreEntity "p01m5mss" as the full hierarchy was not fetched
      */
     public function testRequestingUnfetchedOptionsThrowsException()
     {
