+++ tests/Domain/Entity/SeriesTest.php
@@ -122,7 +122,7 @@ class SeriesTest extends TestCase
 
     /**
      * @expectedException \BBC\ProgrammesPagesService\Domain\Exception\DataNotFetchedException
-     * @expectedExceptionMessage Could not get options of Programme "p01m5mss"as the full hierarchy was not fetched
+     * @expectedExceptionMessage Could not get options of CoreEntity "p01m5mss" as the full hierarchy was not fetched
      */
     public function testRequestingUnfetchedOptionsThrowsException()
     {
